+++ src/error.rs
@@ -156,6 +156,7 @@ win32_errors! {
     ERROR_NO_MORE_ITEMS = 259 => "no more data is available",
     ERROR_INVALID_SID = 1337 => "the security ID structure is invalid",
     ERROR_SERVICE_DOES_NOT_EXIST = 1060 => "the specified service does not exist as an installed service",
+    ERROR_SERVICE_ALREADY_RUNNING = 1056 => "an instance of the service is already running",
 }
 
 impl core::error::Error for Win32Error {}